Fix issues with handling dirty values
`TCOD_sys_console_to_bitmap` can now be given a console to be used as a
cache as a parameter, this cache is always updated with the previous
values.

Code to mark consoles dirty has been moved to `sys_sdl_c.c`.

Global variable used to cache the previous console has been moved to a
static variable in `sys_sdl2_c.c`.