TCOD_console_is_active() now returns keyboard focus.
Previously was using SDL_WINDOWEVENT_MAXIMIZED and
SDL_WINDOWEVENT_MINIMIZED events, but maximized is not the opposite of
minimized. app_is_active would be set to false when the window was
minimized and because libtcod does not support maximizing windows it would
stay false forever.

app_is_active now uses focus gain/loss events instead, and should be
useful now.