Reset the console cache when the tileset is changed.
There were two points of failure to fix:

If the tileset was changed to a different size and the console was the
same size then the render texture would be reset to a blank screen which
would only update as tiles were changed.

If the tileset was changed to another one of the same size then the old
tilesets tiles would be left behind as artifacts.