Add testing and benchmarks for BresenhamLine class.
Switched to the random_access_iterator_tag to let C++ know it can determine
the size of the output array.

Use index_ variable to reliably determine the position of the iterator.

Added noexcept, const, and constexpr keywords.

constexpr is mostly useless since the constructor calls a C function and
can't be contexpr.  This might be fixed later.

Removed const from the orig_ and dest_ variables since it would
delete the default constructors.